requirements.txt
numpy
pandas
matplotlib
Pillow
tqdm
tensorflow
scikit-learn
joblib
seaborn
opencv-python
streamlit